1-Simple_ML_REST/requirements.txt
flask
gevent
requests
pillow